producer/requirements.producer.txt
websockets
pydantic
kafka-python